requirements.txt
Django==1.11.5
Unipath==1.1
django-model-utils==3.0.0
djangorestframework==3.6.3
psycopg2==2.7.3.1
requests
Pillow
django-filter
googlemaps==2.5.1